+++ frontend/styles/home/Attendance.module.css
@@ -1,0 +1,4 @@
+.chartContainer {
+  display: flex;
+  justify-content: center;
+}
