+++ frontend/styles/home/Votings.module.css
@@ -1,0 +1,51 @@
+.voteContainer {
+  padding: 30px 30px;
+  border-radius: 10px;
+  background-color: rgba(0, 0, 0, 0.050);
+  margin-bottom: 20px;
+}
+
+.head {
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+}
+
+.head h3 {
+  font-size: 1.3em;
+  font-weight: bolder;
+}
+
+.date {
+  font-size: 1em;
+}
+
+.matters {
+  font-weight: 400;
+  font-size: 1.2em;
+}
+
+.link {
+  text-align: right;
+}
+
+.link button {
+  background-color: #f5f5f5;
+  border: 1px solid #5f5f5f;
+  border-radius: 5px;
+  padding: 10px 15px;
+  font-size: 1em;
+  font-weight: 600;
+  color: #5f5f5f;
+  cursor: pointer;
+}
+
+.greenVote b {
+  font-size: 1.1em;
+  color: #06c96e;
+}
+
+.redVote b {
+  font-size: 1.1em;
+  color: #f00;
+}
